fix(cli): exec did not correctly forward exit code on Windows

+++ cli/exec.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"log"
 	"net/http"
@@ -382,6 +383,7 @@ func runSubProcess(command string, args []string, env []string) (int, error) {
 
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan)
+	defer signal.Stop(sigChan)
 
 	if err := cmd.Start(); err != nil {
 		return 0, err
@@ -396,13 +398,19 @@ func runSubProcess(command string, args []string, env []string) (int, error) {
 	}()
 
 	if err := cmd.Wait(); err != nil {
+		// Nonzero exit code results in an ExitError from Wait()!
+		var exitErr *osexec.ExitError
+		if errors.As(err, &exitErr) {
+			// Forward exit code to caller
+			return exitErr.ExitCode(), nil
+		}
+
 		_ = cmd.Process.Signal(os.Kill)
-		return 0, fmt.Errorf("Failed to wait for command termination: %v", err)
+		return 0, fmt.Errorf("Failed to wait for command termination: %w", err)
 	}
 
-	waitStatus := cmd.ProcessState.Sys().(syscall.WaitStatus)
-
-	return waitStatus.ExitStatus(), nil
+	// Wait returned nil so exit code should be 0, but return the process exit code explicitly for clarity.
+	return cmd.ProcessState.ExitCode(), nil
 }
 
 func doExecSyscall(command string, args []string, env []string) error {
